Add Terraform language support
As well as adding a Treesitter parser and an LSP for Terraform, this
commit also tweaks the calls to setup for nvim-treesitter and
mason-lspconfig so that they use the ensure_installed options. This
not only means automatically keeping required parsers/LSPs synced across
environments, but also means there is no need to manually write a script
to do so.

As a result, this commit resolves #11 and also resolves #13.